# Reserve vector capacity to avoid repeated allocations